Move nested lib files to src root
