+++ src/browser.ts
@@ -1,7 +1,7 @@
 export * from './types'
 export * from './errors'
-export * from './lib'
-import { MeiliSearch } from './lib/clients/browser-client'
+export * from './indexes'
+import { MeiliSearch } from './clients/browser-client'
 
 export { MeiliSearch }
 export default MeiliSearch
