+++ src/clients/browser-client.ts
@@ -1,5 +1,5 @@
 import { Client } from './client'
-import { Config } from '../../types'
+import { Config } from '../types'
 
 class MeiliSearch extends Client {
   constructor(config: Config) {
