+++ src/index.ts
@@ -1,7 +1,7 @@
 export * from './types'
 export * from './errors'
-export * from './lib'
-import { MeiliSearch } from './lib/clients/node-client'
+export * from './indexes'
+import { MeiliSearch } from './clients/node-client'
 
 export { MeiliSearch }
 export default MeiliSearch
