--- src/lib/index.ts
@@ -1,3 +1,0 @@
-export * from './http-requests'
-export * from './indexes'
-export * from './utils'
